Add files via upload
The Bank Management System is a web-based application and its main motive is to build an organic and optimal interaction system between the elements of banking mechanisms with a view to profit. With the help of this system , users can perform all the tasks like creating an account, deposit amount, check balance, view all account holder details, close an account and modify an account easily. The purpose of this system is to automate basic banking activities to aid a bank clerk’s day to day operations.
The future scope of implementing a banking management system using C++ is vast. As per the available resources, the banking management system can be developed using C++ programming language with different features such as creating an account, depositing and withdrawing amounts, checking balances, viewing account holder details, closing accounts, and modifying accounts. The program can be implemented using classes and file-handling features of C++. The banking system allows banks to keep track of a few records, and therefore, a bank management software program is needed to simplify the process.